more rfks for type-safe provider logic
